📦 chore(firebase.json): add firebase.json configuration file 🔥 chore(nuxt.config.js): comment out nitro preset configuration 📝 chore(package.json): update package name, version, and description 🎨 chore(src/assets/css/tailwind.css): add background color utility class
Added the firebase.json configuration file to the project. Commented out the nitro preset configuration in the nuxt.config.js file. Updated the package name, version, and description in the package.json file. Added a background color utility class in the tailwind.css file.

🎨 style(Default.vue): update navbar layout and add dropdown menu
🔧 chore(Default.vue): remove unused code and comments
The navbar layout in Default.vue has been updated to include a dropdown menu for mobile devices. The dropdown menu is hidden by default and can be toggled by clicking on a button. The menu contains a link to the home page and a commented-out link to a "Uses" page. Unused code and comments related to selecting a theme have been removed.

🔨 refactor(gui.vue): comment out unused import and component
🎨 style(gui.vue): add fade-in animation and tooltip to social links
The unused import and component `GenericPanel` have been commented out to improve code cleanliness. The social links now have a fade-in animation and a tooltip that displays the link name when hovered over.

🎨 style(index.vue): update terminal background color and greeting message
✨ feat(index.vue): add 'gui' command to switch to GUI mode
🐛 fix(index.vue): add underline style to links and fix email link

The terminal component now has a updated background color to 'bg-base-200' for better visibility. The greeting message now includes a heart emoji and provides instructions to type '--help' to see available commands. Additionally, a new 'gui' command has been added to switch to GUI mode. The commands section has been updated to include the 'gui' command. Links in the terminal now have an underline style, and the email link has been fixed to include the underline style as well.

🎨 style(index.vue): add margin-bottom and width to #prompt-input
The #prompt-input element now has a margin-bottom of 2px and a width of 300px. This improves the styling of the input element in the index.vue page.

🔧 chore(tailwind.config.js): add themes to daisyui configuration
The daisyui configuration in tailwind.config.js now includes a list of themes. These themes can be used to customize the appearance of the application using daisyui. The added themes are: light, dark, cupcake, bumblebee, emerald, corporate, synthwave, retro, cyberpunk, valentine, halloween, garden, forest, aqua, lofi, pastel, fantasy, wireframe, black, luxury, dracula, cmyk, autumn, business, acid, lemonade, night, coffee, and winter.