📁 chore(.firebaserc): add .firebaserc file 📁 chore(firebase-hosting-merge.yml): add firebase-hosting-merge.yml workflow 📁 chore(firebase-hosting-pull-request.yml): add firebase-hosting-pull-request.yml workflow Added the .firebaserc file to configure the default Firebase project to "jofftiquez-dev". Also added the firebase-hosting-merge.yml and firebase-hosting-pull-request.yml workflows for deploying to Firebase Hosting on merge and pull requests respectively. These workflows use the Firebase CLI and GitHub Actions to generate and deploy the static site to the "jofftiquez-dev" project.
🔧 chore(firebase.json): update hosting configuration for static files
The hosting configuration in firebase.json has been updated to include additional settings for static files. The "ignore" property now excludes the firebase.json file, hidden files, and the node_modules directory. The "headers" property has been added to specify custom headers for different file types. This includes setting the "Access-Control-Allow-Origin" header for font files, setting the "Cache-Control" header for JavaScript and CSS files, setting the "Cache-Control" header for image and video files, and setting the "Cache-Control" header for the 404.html file.

📝 chore(404.html): add 404 error page
🆕 Add a new file `404.html` to the `output/public` directory. This file contains the HTML markup for a custom 404 error page. The page displays a "Page Not Found" message along with an explanation of why the user is seeing the error and how to modify the page.

📝 chore(index.html): add Firebase SDK initialization code
The commit adds the necessary code to initialize the Firebase SDK in the index.html file. This code initializes various Firebase features such as authentication, database, firestore, functions, messaging, storage, analytics, remoteConfig, and performance. The code also handles the case when the Firebase SDK fails to load and displays an error message in the console.