🚀 chore(package.json): update Nuxt and DaisyUI versions 🖼️ feat(gui.vue): add image and adjust layout in GUI page
The package.json file was updated to use the latest versions of Nuxt (3.6.5) and DaisyUI (3.9.4). This ensures that the project is using the most recent and stable versions of these dependencies.

In the gui.vue file, an image was added to the page using the `<img>` tag. The layout of the GUI page was also adjusted to display the image on the left side and the content on the right side. This improves the visual presentation of the page and enhances the user experience.